+++ src/core/addition/addition.ts
@@ -199,7 +199,47 @@ export const Addition = {
 		})(),
 	} as const,
 	initMechanics() {
+		SOFTCAPS.create('addpower^1', {
+			name: 'addpower^1',
+			fluid: true,
+			start: new Decimal(2).pow(384),
+			exponent: new Decimal(0.75),
+		});
+		SOFTCAPS.create('addpower^2', {
+			name: 'addpower^2',
+			fluid: true,
+			get start() {
+				let base = new Decimal(2).pow(4096);
 
+				if (player.upgrades[43]) base = base.pow(2);
+				return base;
+			},
+			exponent: new Decimal(0.75),
+		});
+		SOFTCAPS.create('addpower^3', {
+			name: 'addpower^3',
+			fluid: true,
+			start: new Decimal('e40000'),
+			exponent: new Decimal(0.5),
+		});
+		SOFTCAPS.create('addpower^4', {
+			name: 'addpower^4',
+			fluid: true,
+			start: new Decimal('ee5'),
+			get exponent() {
+				let base = new Decimal(4);
+				if (player.milestones.cb6) base = base.pow(0.5);
+				return base.pow(-1);
+			},
+			meta: 1,
+		});
+		SOFTCAPS.create('addpower^5', {
+			name: 'addpower^5',
+			fluid: true,
+			start: new Decimal('ee14'),
+			exponent: new Decimal(0.25),
+			meta: 1,
+		});
 	},
 	addpower_gain(bulk = DC.D_1) {
 		let adding = this.gain().mul(bulk);
@@ -213,8 +253,13 @@ export const Addition = {
 			adding = adding.add(1).pow(feature.SingularityGenerator.getSingularityEffect()).sub(1);
 		if (player.buyables[31].gt(0) && Logarithm.logarithm.upgrades_in_dilated.includes('31'))
 			adding = adding.mul(buyables[31].effect(player.buyables[31]));
-		if (adding.gte(softcaps['addpower^1'].start)) adding = adding.div(softcaps['addpower^1'].start).pow(softcaps['addpower^1'].exponent).mul(softcaps['addpower^1'].start)
-		if (adding.gte(softcaps['addpower^2'].start)) adding = adding.div(softcaps['addpower^2'].start).pow(softcaps['addpower^2'].exponent).mul(softcaps['addpower^2'].start)
+		if (player.singularity.stage < 2) {
+			adding = SOFTCAPS.fluidComputed('addpower^1', adding, player.addpower);
+			adding = SOFTCAPS.fluidComputed('addpower^2', adding, player.addpower);
+			adding = SOFTCAPS.fluidComputed('addpower^3', adding, player.addpower);
+			adding = SOFTCAPS.fluidComputed('addpower^4', adding, player.addpower);
+			adding = SOFTCAPS.fluidComputed('addpower^5', adding, player.addpower);
+		}
 		if (CHALLENGE.inChallenge(0, 3)) {
 			adding = adding.mul(predictableRandom(Math.floor(Date.now() / 40)) > 0.5 ? -1 : 1);
 		}
