+++ src/core/mechanic.ts
@@ -21,7 +21,6 @@ import { ORDINAL_BOOSTER } from './ordinal/ordinal-booster.ts';
 import { Hydra } from './hydra/hydra.ts';
 import { Dilute, DiluteUpgrades } from './hydra/dilute.ts';
 import type { Upgrade } from './upgrade.ts';
-import { DC } from '@/core/constants';
 
 const upgrades = {
 	...Successor.upgrades,
@@ -303,10 +302,7 @@ export const SOFTCAPS = {
 		}
 		if (!softcaps[id].fluid) throw new Error('type error');
 		const s = softcaps[id];
-		let base = this.reach(id, existing) ? s.start.mul(existing.div(s.start).root(s.exponent)) : existing
-		base = base.add(getting).pow(s.exponent).mul(s.start)
-		if (base.lt("ee10")) base = base.sub(existing)
-		return base;
+		return overflow_v2(getting, existing, s);
 	},
 	/**
 	 * 基于获取的软上限
@@ -322,114 +318,6 @@ export const SOFTCAPS = {
 	},
 };
 
-		SOFTCAPS.create('number^1', {
-			name: 'number^1',
-			fluid: true,
-			start: new Decimal(2).pow(256),
-			exponent: new Decimal(0.75),
-		});
-		SOFTCAPS.create('number^2', {
-			name: 'number^2',
-			fluid: true,
-			get start() {
-				let base = new Decimal(2).pow(1024);
-
-				if (player.upgrades[43]) base = base.pow(2);
-				return base;
-			},
-			exponent: new Decimal(0.75),
-		});
-		SOFTCAPS.create('number_C1', {
-			name: 'number_C1',
-			fluid: true,
-			start: DC.D_1,
-			exponent: new Decimal(0.5),
-		});
-		SOFTCAPS.create('number^3', {
-			name: 'number^3',
-			fluid: true,
-			start: new Decimal('e20000'),
-			exponent: new Decimal(0.5),
-		});
-		SOFTCAPS.create('number^4', {
-			name: 'number^4',
-			fluid: true,
-			start: new Decimal('ee5'),
-			get exponent() {
-				let base = new Decimal(4);
-				if (player.milestones.cb6) base = base.pow(0.5);
-				return base.pow(-1);
-			},
-			meta: 1,
-		});
-		SOFTCAPS.create('number^5', {
-			name: 'number^5',
-			fluid: true,
-			start: new Decimal('ee20'),
-			get exponent() {
-				return player.milestones.cb14 ? new Decimal(0.2) : new Decimal(0.1);
-			},
-			meta: 1,
-		});
-		SOFTCAPS.create('addpower^1', {
-			name: 'addpower^1',
-			fluid: true,
-			start: new Decimal(2).pow(384),
-			exponent: new Decimal(0.75),
-		});
-		SOFTCAPS.create('addpower^2', {
-			name: 'addpower^2',
-			fluid: true,
-			get start() {
-				let base = new Decimal(2).pow(4096);
-
-				if (player.upgrades[43]) base = base.pow(2);
-				return base;
-			},
-			exponent: new Decimal(0.75),
-		});
-		SOFTCAPS.create('addpower^3', {
-			name: 'addpower^3',
-			fluid: true,
-			start: new Decimal('e40000'),
-			exponent: new Decimal(0.5),
-		});
-		SOFTCAPS.create('addpower^4', {
-			name: 'addpower^4',
-			fluid: true,
-			start: new Decimal('ee5'),
-			get exponent() {
-				let base = new Decimal(4);
-				if (player.milestones.cb6) base = base.pow(0.5);
-				return base.pow(-1);
-			},
-			meta: 1,
-		});
-		SOFTCAPS.create('addpower^5', {
-			name: 'addpower^5',
-			fluid: true,
-			start: new Decimal('ee14'),
-			exponent: new Decimal(0.25),
-			meta: 1,
-		});
-				SOFTCAPS.create('mulpower^1', {
-			name: 'mulpower^1',
-			fluid: true,
-			start: new Decimal('e5e6'),
-			get exponent() {
-				let base = new Decimal(2.5);
-				if (player.upgrades[47]) base = base.pow(wgEffect()[4]);
-				return DC.D_1.div(base);
-			},
-			meta: 1,
-		});
-		SOFTCAPS.create('mulpower^2', {
-			name: 'mulpower^2',
-			fluid: true,
-			start: new Decimal('ee9'),
-			exponent: new Decimal(0.25),
-			meta: 1,
-		});
 type IMilestone = {
 	requirement: Decimal;
 	currency: string;
