+++ src/core/save/testing.ts
@@ -13,5 +13,3 @@ export function isDeveloper() {
 	const devcode = localStorage.getItem('developercode') ?? '';
 	return sha256(devcode) == code2;
 }
-
-export const pubtest = true;
