+++ src/utils/init.ts
@@ -1,5 +1,6 @@
 import { loadSaves, player, restoreBackup } from '@/core/save';
 import { createApp } from 'vue';
+import App from '@/App.vue';
 import VueLatex from 'vatex';
 import { feature } from '@/core/global.ts';
 import { NUMTHEORY } from '@/core/multiplication/numbertheory.ts';
@@ -15,8 +16,6 @@ import { NON_RECURSIVE } from '@/core/nonrecu/index.ts';
 import ModalService from './Modal.ts';
 import { temp } from '@/core/temp-data.ts';
 import { keyboardEventListener, hardResetMiniGame } from '@/core/minigame/index.ts';
-
-import App from '@/App.vue';
 
 export function init() {
 	try {
@@ -44,6 +43,8 @@ export function init() {
 		player.frozen = false;
 		player.run_a_tick_and_froze = false;
 		startGameLoop();
+		const app = createApp(App);
+		app.use(VueLatex).directive('hold', vHold).mount('#app');
 		hotkeys('a', (event) => {
 			event.preventDefault();
 			feature.ADDITION.UIreset();
@@ -69,8 +70,6 @@ export function init() {
 		hotkeys('ArrowRight', keyboardEventListener);
 		//hardResetMiniGame();
 		player.minigame.interact = 0;
-		const app = createApp(App);
-		app.use(VueLatex).directive('hold', vHold).mount('#app');
 	} catch (e) {
 		stopGameLoop();
 		stopSaveLoop();
